+++ 04-backtracking/0046-permutations.cpp
@@ -18,7 +18,7 @@ class Solution {
             return;
         }
 
-        for (int i = 0;i < nums.size(); ++i) {
+        for (int i = 0; i < static_cast<int>(nums.size()); ++i) {
             if (visited[i]) continue;
